Add files via upload
This is a template for just one award page. I decided to put the download link for the award application in the right column; feedback on that decision would be appreciated. Another notable difference between this and the index/faq pages I'd made is that the columns are not aligned with the sides of the container. Instead, there is one empty Bootstrap column on the left and right. Again, give me feedback.